Snapshot of sierra.base.git from commit 9a59021eabfb29ea184ec4388c3df4f1e827b1bc
From repository at sierra-git.sandia.gov:/git/sierra.base.git

At commit:
commit 9a59021eabfb29ea184ec4388c3df4f1e827b1bc
Author: Nick Roehrig <nsroehr@sandia.gov>
Date:   Tue Mar 28 09:22:19 2017 -0600

    fixed

    - why no unused variable errors?

+++ packages/stk/stk_balance/unit_tests/UnitTestLoadBalanceEmptyMeshOnProc.cpp
@@ -78,7 +78,7 @@ void balance_mesh(stk::mesh::BulkData& bulkData)
 
 void test_that_mesh_is_balanced(const stk::mesh::BulkData& stkMeshBulkData, const unsigned gold_num_elements_per_proc)
 {
-    std::vector<unsigned> counts;
+    std::vector<size_t> counts;
     stk::mesh::count_entities(stkMeshBulkData.mesh_meta_data().locally_owned_part(), stkMeshBulkData, counts);
     EXPECT_EQ(gold_num_elements_per_proc, counts[stk::topology::ELEM_RANK]);
 }
