+++ packages/stk/stk_doc_tests/stk_mesh/createFacesEdgesHex.cpp
@@ -82,7 +82,7 @@ namespace
     // ==================================================
     // VERIFICATION
     stk::mesh::Selector allEntities = stkIo.meta_data().universal_part();
-    std::vector<unsigned> entityCounts;
+    std::vector<size_t> entityCounts;
     stk::mesh::count_entities(allEntities, stkIo.bulk_data(), entityCounts);
     EXPECT_EQ( 512u, entityCounts[stk::topology::ELEMENT_RANK]);
     EXPECT_EQ(1728u, entityCounts[stk::topology::FACE_RANK]);
@@ -126,7 +126,7 @@ namespace
     // ==================================================
     // VERIFICATION
     stk::mesh::Selector allEntities = stkIo.meta_data().universal_part();
-    std::vector<unsigned> entityCounts;
+    std::vector<size_t> entityCounts;
     stk::mesh::count_entities(allEntities, stkIo.bulk_data(), entityCounts);
     EXPECT_EQ( 512u, entityCounts[stk::topology::ELEMENT_RANK]);
     EXPECT_EQ(1728u, entityCounts[stk::topology::FACE_RANK]);
@@ -170,7 +170,7 @@ namespace
     // ==================================================
     // VERIFICATION
     stk::mesh::Selector allEntities = stkIo.meta_data().universal_part();
-    std::vector<unsigned> entityCounts;
+    std::vector<size_t> entityCounts;
     stk::mesh::count_entities(allEntities, stkIo.bulk_data(), entityCounts);
     EXPECT_EQ( 512u, entityCounts[stk::topology::ELEMENT_RANK]);
     EXPECT_EQ(1728u, entityCounts[stk::topology::FACE_RANK]);
@@ -277,7 +277,7 @@ namespace
     }
 
     stk::mesh::Selector allEntities = stkIo.meta_data().universal_part();
-    std::vector<unsigned> entityCounts;
+    std::vector<size_t> entityCounts;
     stk::mesh::count_entities(allEntities, stkIo.bulk_data(), entityCounts);
     EXPECT_EQ(1u, entityCounts[stk::topology::ELEMENT_RANK]);
     EXPECT_EQ(6u, entityCounts[stk::topology::FACE_RANK]);
