+++ packages/stk/stk_performance_tests/stk_mesh/NodalFieldPerf.cpp
@@ -98,7 +98,7 @@ void createMetaAndBulkData(stk::io::StkMeshIoBroker &exodusFileReader, stk::mesh
     std::cerr << "Finished Reading Mesh: " << exodusFileName << std::endl;
 
     stk::mesh::Selector allEntities = stkMeshMetaData.universal_part();
-    std::vector<unsigned> entityCounts;
+    std::vector<size_t> entityCounts;
     stk::mesh::count_entities(allEntities, stkMeshBulkData, entityCounts);
     size_t numElements = entityCounts[stk::topology::ELEMENT_RANK];
     size_t numNodes = entityCounts[stk::topology::NODE_RANK];
